Add xml output for match_env

+++ src/feature/co_table.cpp
@@ -59,7 +59,7 @@ bool CoTable::checkCoexist(MatchEnvPtr env_a, MatchEnvPtr env_b,
   RealEnvPtr real_env_b = env_b->getRealEnvPtr();
   for (int i = 0; i < real_env_a->getPeakNum(); i++) {
     int a_idx = real_env_a->getPeakIdx(i);
-    for (size_t j = 0; j < real_env_b->getPeakNum(); j++) {
+    for (int j = 0; j < real_env_b->getPeakNum(); j++) {
       int b_idx = real_env_b->getPeakIdx(j);
       if (real_env_a->isExist(i) && a_idx == b_idx) {
         cnt_share++;
